# schemas/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SWEN Bank API",
    "description": "API Documentation For SWEN Bank",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3030",
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/DebtReminders": {
      "post": {
        "tags": [
          "DebtReminders"
        ],
        "summary": "Create a new debt reminder. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the debt reminder information",
            "required": true,
            "schema": {
              "properties": {
                "account_id": {
                  "type": "integer"
                },
                "amount": {
                  "type": "integer"
                },
                "note": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debt reminder is created",
            "schema": {
              "$ref": "#/definitions/debtReminder"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "DebtReminders"
        ],
        "summary": "Get all debt reminder of an account. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get debt reminder list",
            "schema": {
              "$ref": "#/definitions/debtReminder"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/DebtReminders/{DebtReminders_ID}": {
      "delete": {
        "tags": [
          "DebtReminders"
        ],
        "summary": "Delete a debt reminder. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "DebtReminders_ID",
            "in": "path",
            "required": true,
            "description": "ID of a debt reminder",
            "type": "integer"
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the debt reminder information",
            "required": true,
            "schema": {
              "properties": {
                "note": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debt reminder not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "Debt reminder is deleted",
            "schema": {
              "$ref": "#/definitions/successResponse"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "DebtReminders"
        ],
        "summary": "Update a debt reminder. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "DebtReminders_ID",
            "in": "path",
            "required": true,
            "description": "ID of a debt reminder",
            "type": "integer"
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the debt reminder information",
            "required": true,
            "schema": {
              "properties": {
                "account_id": {
                  "type": "integer"
                },
                "amount": {
                  "type": "integer"
                },
                "note": {
                  "type": "string"
                },
                "isPaid": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debt reminder not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "Debt reminder is updated",
            "schema": {
              "$ref": "#/definitions/debtReminder"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/DebtReminders/{DebtReminders_ID}/Pay": {
      "post": {
        "tags": [
          "DebtReminders"
        ],
        "summary": "Pay for a debt reminder. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "DebtReminders_ID",
            "in": "path",
            "required": true,
            "description": "ID of a debt reminder",
            "type": "integer"
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the information to pay for the debt reminder",
            "required": true,
            "schema": {
              "properties": {
                "source_account_uuid": {
                  "type": "string"
                },
                "destination_account_uuid": {
                  "type": "string"
                },
                "amount": {
                  "type": "integer"
                },
                "note": {
                  "type": "string"
                },
                "fee_is_paid_by_receiver": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debt reminder or bank account is not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "Transaction transfer ID",
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                }
              }
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "debtReminder": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "amount": {
          "type": "integer"
        },
        "note": {
          "type": "string"
        },
        "isPaid": {
          "type": "boolean"
        },
        "user_id": {
          "type": "integer"
        },
        "account_id": {
          "type": "integer"
        },
        "created_at": {
          "type": "string"
        }
      }
    },
...
    "errorResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "The error message"
        }
      }
    },
    "successResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "The success message"
        }
      }
    }
  }
}